orders/subscriptions.py
"""
GraphQL subscriptions for real-time order status updates.
"""
import graphene
from channels.layers import get_channel_layer
from asgiref.sync import async_to_sync

from .models import Order, OrderEvent


class Subscriptions(graphene.ObjectType):
    """
    Root subscription object.
    Note: For full WebSocket functionality, implement using Django Channels consumers.
    """
    # Placeholder subscriptions - actual implementation via WebSocket consumers
    # Using String type to avoid circular imports and duplicate type registration
    order_status_updated = graphene.String()
    order_event = graphene.String()
    user_order_update = graphene.String()

    def resolve_order_status_updated(self, info):
        # This is a placeholder - actual subscription logic handled by consumers
        return "Subscription placeholder"

    def resolve_order_event(self, info):
        # This is a placeholder - actual subscription logic handled by consumers  
        return "Subscription placeholder"

    def resolve_user_order_update(self, info):
        # This is a placeholder - actual subscription logic handled by consumers
        return "Subscription placeholder"


def send_order_status_update(order_id, status, message=None):
    """
    Send order status update to WebSocket subscribers.
    """
    channel_layer = get_channel_layer()
    
    async_to_sync(channel_layer.group_send)(
        f"order_{order_id}",
        {
            "type": "order_status_update",
            "order_id": str(order_id),
            "status": status,
            "message": message or "",
            "timestamp": None  # Will be set by consumer
        }
    )


def send_order_event(order_id, event_type, message):
    """
    Send order event to WebSocket subscribers.
    """
    channel_layer = get_channel_layer()
    
    async_to_sync(channel_layer.group_send)(
        f"order_{order_id}",
        {
            "type": "order_event",
            "order_id": str(order_id),
            "event_type": event_type,
            "message": message,
            "timestamp": None  # Will be set by consumer
        }
    )


def send_user_order_update(user_id, order_id, status, message=None):
    """
    Send order update to user's general order subscription.
    """
    channel_layer = get_channel_layer()
    
    async_to_sync(channel_layer.group_send)(
        f"user_orders_{user_id}",
        {
            "type": "order_update",
            "order_id": str(order_id),
            "status": status,
            "message": message or "",
            "timestamp": None  # Will be set by consumer
        }
    )


def send_new_order_notification(user_id, order_id, order_number, total_amount):
    """
    Send new order notification to user.
    """
    channel_layer = get_channel_layer()
    
    async_to_sync(channel_layer.group_send)(
        f"user_orders_{user_id}",
        {
            "type": "new_order",
            "order_id": str(order_id),
            "order_number": order_number,
            "total_amount": str(total_amount),
            "timestamp": None  # Will be set by consumer
        }
    ) 