requirements.txt
Django==4.2.7
django-environ==0.11.2
graphene-django==3.0.0
graphene==3.3
django-graphql-jwt==0.4.0
django-cors-headers==4.3.1
django-filter==23.5
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
PyJWT==2.8.0
cryptography==41.0.7